DSBP-tabs initial setup.
