+++ src/components/DSBP/CustomizeView/index.scss
@@ -41,10 +41,6 @@
     overflow-y: auto;
   }
 
-  .cursorDefault {
-    cursor: default;
-  }
-
   .p-dialog-content {
     padding: 0 32px 20px 32px !important;
     border-radius: 0 0 20px 20px;
