+++ src/components/DSBP/PMPSpecificTabView/index.scss
@@ -1,0 +1,6 @@
+.tabsCompoMain {
+    background-color: #eaecea;
+    border-radius: 24px 24px 0 0;
+    height: 400px;
+    overflow-y: auto;
+}
